KASM-4323 KASM-4320 Layering and Locales

+++ src/ubuntu/install/kali/install_kali.sh
@@ -25,8 +25,9 @@ apt-get purge -y \
 rm -f /usr/share/xfce4/panel/plugins/power-manager-plugin.desktop
 
 # Cleanup
-rm -rf \
-  /var/lib/apt/lists/* \
-  /var/tmp/* \
-  /tmp/*
-
+if [ -z ${SKIP_CLEAN+x} ]; then
+  apt-get autoclean
+  rm -rf \
+    /var/lib/apt/lists/* \
+    /var/tmp/*
+fi
