+++ src/ubuntu/install/retroarch/install_retroarch.sh
@@ -38,8 +38,9 @@ EOL
 chmod +x /usr/bin/desktop_ready
 
 # Cleanup
-apt-get autoclean
-rm -rf \
+if [ -z ${SKIP_CLEAN+x} ]; then
+  apt-get autoclean
+  rm -rf \
     /var/lib/apt/lists/* \
-    /var/tmp/* \
-    /tmp/*
+    /var/tmp/*
+fi
